requirements.txt
streamlit
crewai
langchain
pypdf
openai
reportlab
python-dotenv
fpdf 
weasyprint
markdown
